Make testing easier
Uses a mockable clock so that we can build tests with that to make it
easier to deal with tasks in the future.

Adds some atomic checks around current cost and item count although we
don't really care too much if we exceed either cost or total count as
long as we're in the same ballpark.